# docs/api-docs.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Auth","description":"User authorization"}],"paths":{"/api/comments/{id}":{"get":{"tags":["Comments"],"summary":"Get comment","description":"Returns a comment with the specified ID","operationId":"getCommentById","parameters":[],"responses":{"200":{"description":"Comment returned","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentResponseDTO"}}}},"404":{"description":"Comment not found"}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Comments"],"summary":"Edit comment","description":"Edits a comment with the specified ID","operationId":"editCommentById","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequestDTO"}}},"required":true},"responses":{"200":{"description":"Comment edited"},"404":{"description":"Comment not found"}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Comments"],"summary":"Delete comment","description":"Deletes a comment with the specified ID","operationId":"deleteCommentById","parameters":[],"responses":{"204":{"description":"Comment deleted"},"404":{"description":"Comment not found"}},"security":[{"bearerAuth":[]}]}},"/api/blog-posts/{id}":{"get":{"tags":["Blogposts"],"summary":"Get blogpost","description":"Returns a blogpost with the specified ID","operationId":"getBlogPostById","parameters":[],"responses":{"200":{"description":"Blogpost returned","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BlogPostResponseDTO"}}}},"404":{"description":"Blogpost not found"}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Blogposts"],"summary":"Edit blogpost","description":"Edits a blogpost with the specified ID","operationId":"editBlogPostById","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogPostRequestDTO"}}},"required":true},"responses":{"200":{"description":"Blogpost edited"},"404":{"description":"Blogpost not found"}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Blogposts"],"summary":"Delete blogpost","description":"Deletes a blogpost with the specified ID","operationId":"deleteBlogPostById","parameters":[],"responses":{"204":{"description":"Blogpost deleted"},"404":{"description":"Blogpost not found"}},"security":[{"bearerAuth":[]}]}},"/api/blog-posts":{"get":{"tags":["Blogposts"],"summary":"Get all blogposts","description":"Returns a list of all blogposts","operationId":"getBlogPosts","responses":{"200":{"description":"List of all blogposts","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BlogPostResponseDTO"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Blogposts"],"summary":"Create blogpost","description":"Creates a blogpost","operationId":"createBlogPost","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogPostRequestDTO"}}},"required":true},"responses":{"201":{"description":"Blogpost created"}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Blogposts"],"summary":"Bulk delete blogposts","description":"Deletes multiple blogposts with the specified IDs","operationId":"bulkDeleteBlogPosts","parameters":[{"name":"ids","in":"query","description":"Comma-separated list of blogpost IDs","required":true,"schema":{"type":"array","items":{"type":"integer","format":"int64"}}}],"responses":{"204":{"description":"All blogposts deleted"},"404":{"description":"One or more blogposts not found"}},"security":[{"bearerAuth":[]}]}},"/api/blog-posts/{id}/comments":{"get":{"tags":["Comments"],"summary":"Get comments for blogpost","description":"Returns a list of all comments created on a certain blogpost","operationId":"getCommentsByPostId","parameters":[],"responses":{"200":{"description":"Comments on blogpost returned","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponseDTO"}}}}},"404":{"description":"Blogpost not found"}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Comments"],"summary":"Create comment","description":"Creates a comment on the specified blog post ID","operationId":"createComment","parameters":[],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequestDTO"}}},"required":true},"responses":{"201":{"description":"Comments created"},"404":{"description":"Blogpost not found"}},"security":[{"bearerAuth":[]}]}},"/api/auth/register":{"post":{"tags":["Auth"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUserDTO"}}},"required":true},"responses":{"201":{"description":"User registered and token returned","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/auth/login":{"post":{"tags":["Auth"],"operationId":"loginUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUserDTO"}}},"required":true},"responses":{"200":{"description":"Valid credentials passed and token returned","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Invalid credentials passed"}}}},"/api/users/{id}":{"get":{"tags":["Users"],"summary":"Get user by id","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"User returned","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PublicUserDTO"}}}},"404":{"description":"User not found"}},"security":[{"bearerAuth":[]}]}},"/api/users/{id}/posts":{"get":{"tags":["Blogposts"],"summary":"Get all blogposts by user","description":"Get all the blogposts created by the specified user","operationId":"getBlogPostsByUserId","parameters":[],"responses":{"200":{"description":"List of blogposts created by the user","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BlogPostResponseDTO"}}}}},"404":{"description":"User not found"}},"security":[{"bearerAuth":[]}]}},"/api/users/{id}/comments":{"get":{"tags":["Comments"],"summary":"Get all comments by user","description":"Returns a list of all comments created by a certain user","operationId":"getCommentsByAuthorId","parameters":[],"responses":{"200":{"description":"Comments by user returned","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CommentResponseDTO"}}}}},"404":{"description":"User not found"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"CommentRequestDTO":{"type":"object","properties":{"body":{"type":"string","maxLength":300,"minLength":0},"parentCommentId":{"type":"integer","format":"int64"}},"required":["body"]},"BlogPostRequestDTO":{"type":"object","properties":{"title":{"type":"string","maxLength":255,"minLength":0},"body":{"type":"string","maxLength":16000,"minLength":0}},"required":["body","title"]},"AuthUserDTO":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","maxLength":128,"minLength":6}},"required":["password","username"]},"PublicUserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"}}},"BlogPostResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"author":{"$ref":"#/components/schemas/PublicUserDTO"},"title":{"type":"string"},"body":{"type":"string"},"createTime":{"type":"string","format":"date-time"}}},"CommentResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"postId":{"type":"integer","format":"int64"},"parentCommentId":{"type":"integer","format":"int64"},"author":{"$ref":"#/components/schemas/PublicUserDTO"},"body":{"type":"string"}}}}}}